Simplify browser update check handling
